+++ packages/layout/tests/node/setPositionType.test.js
@@ -1,4 +1,4 @@
-import Yoga from '@react-pdf/yoga';
+import Yoga from '@garrettcharliehill/yoga';
 
 import setPositionType from '../../src/node/setPositionType';