+++ packages/renderer/src/index.js
@@ -1,7 +1,7 @@
 import FontStore from '@react-pdf/font';
 import renderPDF from '@react-pdf/render';
 import PDFDocument from '@react-pdf/pdfkit';
-import layoutDocument from '@react-pdf/layout';
+import layoutDocument from '@garrettcharliehill/layout';
 
 import createRenderer from './renderer';
 import { version } from '../package.json';